requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
av==14.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cos-python-sdk-v5==1.9.34
crcmod==1.7
cycler==0.12.1
datasets==3.2.0
decorator==5.2.1
dill==0.3.8
docstring_parser==0.16
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gradio==5.18.0
gradio_client==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
-e git+https://gitclone.com/github.com/hiyouga/LLaMA-Factory.git@820ed764c47ede8f6ae56da89dc48aadc1b000ba#egg=llamafactory
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
modelscope==1.23.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peewee==3.17.9
peft==0.12.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.3.0
protobuf==6.30.0
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge-chinese==1.0.3
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.2.1
starlette==0.46.1
swanboard==0.1.7b1
swankit==0.1.3
swanlab==0.4.9
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.13.3
tomlkit==0.13.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.31.0
triton==3.1.0
trl==0.9.6
typer==0.15.2
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
